math: cast RAND_MAX to float in frand()
Silences clang -Wimplicit-const-int-float-conversion (RAND_MAX is
0x7fffffff, not exactly representable as float).

+++ src/math.hpp
@@ -73,7 +73,7 @@ T mid (const T& a, const T& b, const T& c)
 inline
 float frand()
 {
-  return static_cast<float>(rand()) / RAND_MAX;
+  return static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
 }
 
 inline
